refactor: split overworld_writer.rs into pipeline-step submodules
Break the 1740-line overworld_writer.rs into an overworld_writer/ module
directory mirroring the writer's pipeline steps: types (shared assignment
structs), assign (step 1), grid (step 2), pointers (steps 3/5),
fortress_fx (step 4), metatiles, sprites, and tests. mod.rs keeps the
write_overworld entry point and wires the steps.

Pure code movement: byte-for-byte identical ROM output across seeds and
the full flag matrix; all tests pass, clippy clean.
